+++ src/components/datagrid/datagrid.formatters.js
@@ -625,18 +625,19 @@ const formatters = {
   Button(row, cell, value, col, item, api) {
     let text;
     const disableAttr = isColumnDisabled(row, cell, value, col, item) ? ' disabled' : '';
+    const contentWidth = col.contentWidth ? ` style="width: ${col.contentWidth}"` : '';
+
     if (col.text) {
       text = col.text;
     } else {
       text = (value === null || value === undefined || value === '') ? '' : value.toString();
     }
-    let markup = `<button type="button"${disableAttr} class="${(col.icon ? 'btn-icon' : 'btn-secondary')} row-btn ${(col.cssClass ? col.cssClass : '')}"${(!api.settings.rowNavigation ? '' : ' tabindex="-1"')} >`;
+    let markup = `<button type="button"${disableAttr}${contentWidth} class="${(col.icon ? 'btn-icon' : 'btn-secondary')} row-btn ${(col.cssClass ? col.cssClass : '')}"${(!api.settings.rowNavigation ? '' : ' tabindex="-1"')} >`;
 
     if (col.icon) {
       markup += $.createIcon({ icon: col.icon, file: col.iconFile });
     }
     markup += `<span>${text}</span></button>`;
-
     return markup;
   },